initial CSS and resume updates
The CSS file now holds a stylesheet for us to use, BUT WILL NEED CHANGES SO WE AREN'T STEALING. Also the Pdf reader is now resized to a manageable size and the pdf is altered so it is in one page. Might revert due to new pdf being over twice the size.